Run the water balance on a timer that actually fires
The cycle that reads the weather and grows the deficit was meant to run
every ten minutes. It never did: a DataUpdateCoordinator arms its timer
only while something listens to it, and nothing here does - the sensors
read the store and refresh on a dispatcher signal. After the first
refresh at setup, no further update ever happened, so the balance only
moved when Home Assistant restarted.

The real damage is that the one refresh lands during startup, when the
weather entity usually does not exist yet: the reading comes back empty,
the day keeps no Tmin/Tmax, and with no temperature range ET0 is zero
until midnight. On 29 July, with a measured 22.7/34.2 degrees and no
rain, it recorded 0.00 mm instead of 5.41 - 4.4 mm the lawn lost and the
balance never saw. The previous day was 26% low for the same reason: the
extremes were sampled only in the few moments the cycle happened to run.

The cadence is now an explicit interval owned by the integration, and
the first pass waits for Home Assistant to have started so the weather
entities are there.

Also renames logbook.py to activity_log.py: that filename is reserved -
Home Assistant imports <integration>/logbook.py as a logbook platform
and looks for async_describe_events, raising an AttributeError on every
boot. And clears five linter findings that were failing CI.

+++ custom_components/irrigazione_smart/panel.py
@@ -26,6 +26,7 @@ from homeassistant.helpers.storage import STORAGE_DIR
 from homeassistant.loader import async_get_integration
 from homeassistant.util import dt as dt_util
 
+from .activity_log import get_log
 from .const import (
     CATEGORY_ICONS,
     CATEGORY_LABELS,
@@ -54,7 +55,6 @@ from .hydro import (
     schedule_sequence,
     window_quality,
 )
-from .logbook import get_log
 from .notifier import HOOK_LABELS, build_context, get_notifier
 from .scheduler import get_scheduler
 from .store import WEEKDAYS, IrrigazioneStore, ulid_now
@@ -339,7 +339,7 @@ def _flow_payload(hass: HomeAssistant, system: dict[str, Any]) -> dict[str, Any]
 
 def _hhmm(minute: float) -> str:
     """Minuti dalla mezzanotte in 'HH:MM', anche oltre le 24."""
-    m = int(round(minute)) % (24 * 60)
+    m = round(minute) % (24 * 60)
     return f"{m // 60:02d}:{m % 60:02d}"
